third_party/blink/web_tests/platform/mac-mac10.10/external/wpt/webxr/render_state_vertical_fov_inline.https-expected.txt

This is a testharness.js-based test.
PASS inlineVerticalFieldOfView is set appropriately on inline sessions
Harness: the test ran to completion.
